EXOPS/requirements.txt
# Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
jinja2>=3.1.0

# Auth & Security
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# vSphere
pyVmomi>=8.0.0

# Settings & Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Logging & Utils
loguru>=0.7.0
httpx>=0.27.0

# Dev tools
ruff>=0.4.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
